moved FacebookPost.cs class to its own model Moved puppeteer extensions to Extensions folder
